# Conflicts:
#	platforms/android/assets/www/img/loading.gif
#	platforms/android/assets/www/js/imageTags.js
#	platforms/browser/www/img/loading.gif
#	platforms/browser/www/js/imageTags.js
#	platforms/ios/www/img/loading.gif
#	platforms/ios/www/index.html
#	platforms/ios/www/js/imageTags.js
#	platforms/ios/www/js/imageToText.js
#	platforms/ios/www/js/main.js
#	www/js/imageToText.js

+++ platforms/android/assets/www/js/main.js
@@ -2,7 +2,7 @@
     var familyid = "";
 
     function signInToServer(name){
-         name = '{"familyname" : "myFamily1"}';
+         name = '{"familyName" : "myFamily1"}';
          $.ajax({
          type: "POST",
          url: url + "/login",
@@ -17,7 +17,7 @@
          },
          dataType: "json"
       });
-      setInterval(refreshFeed,5000);
+      setInterval(refreshFeed,30000);
     }
 
     function refreshFeed(){
@@ -27,20 +27,33 @@
                 crossDomain: true,
                 contentType: "json",
                 success: function (result) {
-                    updateLocalStorage(result);
+                    if(result != null){
+                        updateLocalStorage(result);
+                    }else{
+                        loadToDoList();
+                    }
                 },
                 error: function(xhr,status,error){
+                        loadToDoList();                    
                 },
                 dataType: "json"
             });
         }
 
         function updateLocalStorage(result){
+            var theList = JSON.parse(window.localStorage.getItem("todoList"));
+            var check = 0;
             if(result != null && result != undefined && result.fridgeList != null){
                 fridgeList = result.fridgeList;
                 var todoArray = {};
                 for(var i = 0; i < fridgeList.length; i++){
+                    if(theList != null && theList[i]!= null && theList[i].check != null ){
+                        check = theList[i].check;
+                    }else{
+                        check = 0;
+                    }
                     todoArray["row" + i] = {
+                        check: check,
                         text:fridgeList[i].product.name,
                         Quantity:fridgeList[i].quantity,
                         barcode : fridgeList[i].product.barcode
@@ -49,12 +62,18 @@
                 window.localStorage.setItem("todoList", JSON.stringify(todoArray));
             }
 
+            theList = JSON.parse(window.localStorage.getItem("productList"));
             if(result != null && result != undefined && result.productList != null){
                 products = result.productList;
                 var productListArray = {};
                 for(var i = 0; i < products.length; i++){
+                    if(theList != null && theList[i]!= null && theList[i].check != null ){
+                        check = theList[i].check;
+                    }else{
+                        check = 0;
+                    }
                     productListArray["row" + i] = {
-                        check: false,
+                        check: check,
                         text: products[i].name,
                         Quantity: products[i].totalConsumed,
                         barcode : products[i].barcode
@@ -63,12 +82,18 @@
                 window.localStorage.setItem("productList", JSON.stringify(productListArray));
             }
 
+            theList = JSON.parse(window.localStorage.getItem("garbageList"));
             if(result != null && result != undefined && result.garbageList != null){
                 garbages = result.garbageList;
                 var garbageListArray = {};
                 for(var i = 0; i < garbages.length; i++){
+                    if(theList != null && theList[i]!= null && theList[i].check != null ){
+                        check = theList[i].check;
+                    }else{
+                        check = 0;
+                    }
                     garbageListArray["row" + i] = {
-                        check:false,
+                        check:check,
                         text: garbages[i].product.name,
                         Quantity: garbages[i].quantity,
                         barcode : garbages[i].product.barcode
@@ -85,25 +110,16 @@
     function toggle_visibility(id1,id2,id3) {
             $("div[id^='Section']").hide();
             $("#" + id1).show();
-            }
+    }
 
-            document.addEventListener('deviceready', function () {
-                // basic usage
-                TTS
-                    .speak('I have Successfully added ', function () {
-                    }, function (reason) {
-                    });
-            }, false);
-
-
-            function scan(){
-                cordova.plugins.barcodeScanner.scan(function(result){
-                    addItemToKitchen(result.text);
-                    speak(result.text);
-                },function(error){
-                    alert(JSON.stringify(error));
-                });
-            }
+        function scan(){
+            cordova.plugins.barcodeScanner.scan(function(result){
+                addItemToKitchen(result.text, "dataTable1",true);
+                speak(result.text);
+            },function(error){
+                alert(JSON.stringify(error));
+            });
+        }
 
         function createNewToDo(div) {
             var kitchenList = {};
@@ -125,20 +141,174 @@
             }
         }
 
-        function addItemToKitchen(todo,div){
+        function addItemToKitchen(todo,div,isbar){
             kitchenList = {
                     check: 0,
                     text: todo,
                     Quantity: 1,
                     barcode:""
             };
-            addTableRow(kitchenList, false, div);
+            if(div != null && div != ""){
+			    addToServer(kitchenList,"");
+            }
             speak(todo);
         }
 
+        function addItemToKitchenWithOCR(todo,bar){
+        var data = {
+				itemList: [
+				{
+					name: todo,
+                    barcode: bar,
+                    quantity:1
+				}
+				]
+			}
+            alert(JSON.stringify(data));
+			$.ajax({
+				type: "POST",
+				url: url + "/family/" + familyid + "/addToFridge",
+                data: JSON.stringify(data),
+				crossDomain: true,
+				dataType: "json",
+				success: function (result) {
+				},
+				error: function(xhr,status,error){
+				},
+				dataType: "json"
+			});
+        }
+
+
+
+		function generateData(res,str) {
+			var result = res;
+			data = {
+				itemList: [
+				{
+					name: result.text
+				}
+				]
+			}
+			return data;
+		}
+
+        function generateDataForModification(res,str,quantity) {
+			data = {
+				itemList: [
+				{
+					barcode: str,
+                    name: res,
+                    quantity:quantity
+				}
+				]
+			}
+			return data;
+		}
+
+        function modifyOnServer(cell){
+            var str = $(cell).parent().parent().find("td:first").find('input')[0].name;
+            var res = $(cell).parent().parent().find("td:eq(1)").find('label')[0].text;
+            var quantity = $(cell).parent().parent().find("td:eq(2)").find('input')[0].value;
+
+            console.log($(cell).parent().parent().find("td:first").find('input'));
+            
+            if(quantity == null || quantity == ""){
+                quantity = 1;
+            }
+            var data = generateDataForModification(res,str,quantity);
+            $.ajax({
+				type: "POST",
+				url: url + "/family/" + familyid + "/addToFridge",
+                data: data,
+				crossDomain: true,
+				dataType: "json",
+				success: function (result) {
+				},
+				error: function(xhr,status,error){
+				},
+				dataType: "json"
+			});
+        }
+
+        function generateDataForShopping(res,str) {
+			var result = res;
+            var data = "";
+            var d= "";
+            if(res != null){
+                var outer ='{"itemList": ['; 
+                for(i =0 ; i < res.length-1;i++){
+                    data += '{"barcode" : "' + res[i] + '"},';
+                }
+                data += '{"barcode" :"'  + res[res.length-1] + '"}'
+                var later = ' ]}';
+                d = outer + data + later;
+            }      
+			return d;
+		}
+    
+        function removeFromServer(){
+            var res = getAllCheckedBarCodes("dataTable1")
+            var data = generateDataForShopping(res);
+            $.ajax({
+            type: "POST",
+            url: url + "/family/" + familyid + "/addToGarbage",
+            data: data,
+            crossDomain: false,
+            contentType: "application/json",
+            success: function (result) {
+                updateLocalStorage(result);
+            },
+            error: function(xhr,status,error){
+            },
+            dataType: "json"
+        });
+    }
+
+        function getAllCheckedBarCodes(div){
+            var selected = [];
+            $('#' + div + ' input:checked').each(function() {
+                selected.push($(this).attr('name'));
+            });
+            console.log(selected);
+            return selected;
+        }
+
+
+        function generateDataForMove(res,str) {
+            var result = res;
+            alert(JSON.stringify(result));
+            data = {
+                "itemList": [
+                {
+                    "barcode": result.text
+                }
+                ]
+            }
+        }
+		
+		function addToServer(res,str){
+			var data = generateData(res,str);
+            $("#loader").show();
+			$.ajax({
+				type: "POST",
+				url: url + "/family/" + familyid + "/addToFridge",
+				data: data,
+				crossDomain: true,
+				dataType: "json",
+				success: function (result) {
+                    updateLocalStorage(result);
+                    $("#loader").hide();
+				},
+				error: function(xhr,status,error){
+				},
+				dataType: "json"
+			});
+		}
+        
         var rowID = 0;
 
-        function addTableRow(kitchenList, appIsLoading, div) {
+        function addTableRow(kitchenList, appIsLoading, div, isbar) {
             rowID += 1;
             var table = document.getElementById(div);
             var rowCount = table.rows.length;
@@ -147,27 +317,39 @@
             var cell1 = row.insertCell(0);
             var element1 = document.createElement("input");
             element1.type = "checkbox";
-            element1.name = "chkbox[]";
+            element1.name = kitchenList['barcode'];
             element1.checked = kitchenList["check"];
-            element1.setAttribute("onclick", "checkboxClicked()");
+            element1.id = "check" + rowID;
+            element1.setAttribute("onclick", "checkboxClicked('"+ div + "')");
             cell1.style = "width:20% !important";
             cell1.appendChild(element1);
 
-
             var cell2 = row.insertCell(1);
             var element2 = document.createElement("Label");
-            element2.setAttribute("for",kitchenList["text"]);
-            element2.innerHTML = kitchenList["text"];
+            element2.setAttribute("for",kitchenList["barcode"]);
+            var shouldAdd = false;
+            if(kitchenList["text"] == null || kitchenList["text"] == "" || (isbar !=null && isbar == true)){
+                element2.innerHTML = kitchenList["barcode"];
+                var btn = document.createElement("BUTTON");
+                btn.innerHTML = "Use OCR";
+                btn.setAttribute("onclick", "getPhoto(pictureSource.CAMERA, '" + kitchenList['barcode'] + "');");
+                shouldAdd = true;
+            }else{
+                element2.innerHTML = kitchenList["text"];
+            }
             element2.name = "txtbox[]";
             element2.class = "contenttd";
             element2.id = "text" + rowID;
-            element2.setAttribute("onchange", "saveToDoList("+ div + ")");
+            element2.setAttribute("onchange", "saveToDoList('"+ div + "')");
             cell2.style = "width:50% !important";
             cell2.appendChild(element2);
+            if(shouldAdd){
+                cell2.appendChild(btn);
+            }
+            
 
             var cell3 = row.insertCell(2);
             var element3 = document.createElement("input");
-            console.log("todo " + kitchenList["Quantity"]);
             if(kitchenList["Quantity"] == null || kitchenList["Quantity"] == undefined){
                 element3.setAttribute("value",1);
             }else{
@@ -175,7 +357,7 @@
             }
             element3.name = "txtbox[]";
             element3.id = "Quantity" + rowID;
-            element3.setAttribute("onchange", "saveToDoList("+ div + ")");
+            element3.setAttribute("onblur", "modifyOnServer(this)");
             element3.style = "width:50%";
             cell3.style = "width:30% !important";
             cell3.appendChild(element3);
@@ -248,7 +430,6 @@
                 }
             }
             saveToDoList(div);
-            alert("Completed Tasks Were Removed Successfully.");
         }
 
 
@@ -260,7 +441,7 @@
 
             var table = document.getElementById(div);
             var rowCount = table.rows.length;
-
+            console.log(rowCount);
             if (rowCount != 0) {
                 // loop through all rows of the table
                 for (var i = 0; i < rowCount; i++) {
@@ -275,12 +456,14 @@
                     }
 
                     // retrieve the content of the to-do
-                    var textbox = row.cells[1].childNodes[0];
-                    textValue = textbox.innerHTML;
+                    // var textbox = row.cells[1].childNodes[0];
+                    // textValue = textbox.innerHTML;
+                    // todoArray = window.localStorage.getItem("todoList");
+                    // var barcode = todoArray[i].barcode;
+                    // console.log(quantValue);
 
-                    var qunatText = row.cells[2].childNodes[0];
-                    quantValue = qunatText.value;
-
+                    var quantText = row.cells[2].childNodes[0];
+                    quantValue = quantText.value;
                     // populate the array
                     todoArray["row" + i] = {
                         check: checkBoxState,
@@ -291,8 +474,6 @@
             } else {
                 todoArray = null;
             }
-
-            // use the local storage API to persist the data as JSON
             window.localStorage.setItem("todoList", JSON.stringify(todoArray));
         }
 
@@ -307,10 +488,11 @@
                 for (var obj in theList) {
                     count++;
                 }
-
-                deleteAllRows("dataTable1");
-                for (var i = 0; i < count; i++) {
-                    addTableRow(theList["row" + i], true, "dataTable1");
+                if(theList != null){
+                    deleteAllRows("dataTable1");
+                    for (var i = 0; i < count; i++) {
+                        addTableRow(theList["row" + i], true, "dataTable1");
+                    }
                 }
             }
         }
@@ -325,15 +507,15 @@
                 for (var obj in theList) {
                     count++;
                 }
-
-                deleteAllRows("dataTable2");
-                for (var i = 0; i < count; i++) {
-                    addTableRow(theList["row" + i], true, "dataTable2");
+                if(theList != null){
+                    deleteAllRows("dataTable2");
+                    for (var i = 0; i < count; i++) {
+                        addTableRow(theList["row" + i], true, "dataTable2");
+                    }
                 }
             }
         }
 
-    // load the to-do list
             function loadProductList() {
                 var theList = JSON.parse(window.localStorage.getItem("productList"));
                 if (null == theList || theList == "null") {
@@ -343,11 +525,12 @@
                     for (var obj in theList) {
                         count++;
                     }
-
+                if(theList != null){
                     deleteAllRows("dataTable3");
                     for (var i = 0; i < count; i++) {
                         addTableRow(theList["row" + i], true, "dataTable3");
                     }
+                }
                 }
             }
 
@@ -362,6 +545,7 @@
             }
             saveToDoList(div);
         }
+
 
         function toggle_sidebar()
         {
@@ -410,22 +594,22 @@
                 }
             }
 
-    function addLang(loc, lang) {
-        var langs = document.getElementById('langs');
-        var langOption = document.createElement("OPTION") 
-        langOption.innerText = lang; 
-        langOption.value = loc;
-            langs.options.add(langOption);
-    }
-    
-    function changeLang() {
-        var yourSelect = document.getElementById('langs');
-            window.plugins.tts.setLanguage(yourSelect.options[yourSelect.selectedIndex].value, win, fail);
-    }
-    
-    function win(result) {
-        console.log(result);
-    }
+        function addLang(loc, lang) {
+            var langs = document.getElementById('langs');
+            var langOption = document.createElement("OPTION") 
+            langOption.innerText = lang; 
+            langOption.value = loc;
+                langs.options.add(langOption);
+        }
+        
+        function changeLang() {
+            var yourSelect = document.getElementById('langs');
+                window.plugins.tts.setLanguage(yourSelect.options[yourSelect.selectedIndex].value, win, fail);
+        }
+        
+        function win(result) {
+            console.log(result);
+        }
         
         function fail(result) {
             console.log("Error = " + result);
@@ -437,12 +621,12 @@
         }
 
         function voicerec() {
-        var maxMatches = 1;
-        var promptString = "Speak now"; // optional
-        var language = "en-US";         // optional
-        window.plugins.speechrecognizer.startRecognize(function(result){
-        alert(result);
-        }, function(errorMessage){
-            console.log("Error message: " + errorMessage);
-        }, maxMatches, promptString, language);
+            var maxMatches = 1;
+            var promptString = "Speak now"; // optional
+            var language = "en-US";         // optional
+            window.plugins.speechrecognizer.startRecognize(function(result){
+            alert(result);
+            }, function(errorMessage){
+                console.log("Error message: " + errorMessage);
+            }, maxMatches, promptString, language);
     }
